@@ -11,16 +11,21 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 //===----------------------------------------------------------------------===//
+import HTTPTypes
 
-import class Foundation.JSONEncoder
+#if canImport(FoundationEssentials)
+import FoundationEssentials
+#else
+import Foundation
+#endif
 
 // https://docs.aws.amazon.com/lambda/latest/dg/urls-invocation.html
 
 /// FunctionURLRequest contains data coming from a bare Lambda Function URL
-public struct FunctionURLRequest: Codable {
-    public struct Context: Codable {
-        public struct Authorizer: Codable {
-            public struct IAMAuthorizer: Codable {
+public struct FunctionURLRequest: Codable, Sendable {
+    public struct Context: Codable, Sendable {
+        public struct Authorizer: Codable, Sendable {
+            public struct IAMAuthorizer: Codable, Sendable {
                 public let accessKey: String
 
                 public let accountId: String
@@ -36,8 +41,8 @@ public struct FunctionURLRequest: Codable {
             public let iam: IAMAuthorizer?
         }
 
-        public struct HTTP: Codable {
-            public let method: HTTPMethod
+        public struct HTTP: Codable, Sendable {
+            public let method: HTTPRequest.Method
             public let path: String
             public let `protocol`: String
             public let sourceIp: String
@@ -80,15 +85,15 @@ public struct FunctionURLRequest: Codable {
 
 // MARK: - Response -
 
-public struct FunctionURLResponse: Codable {
-    public var statusCode: HTTPResponseStatus
+public struct FunctionURLResponse: Codable, Sendable {
+    public var statusCode: HTTPResponse.Status
     public var headers: HTTPHeaders?
     public var body: String?
     public let cookies: [String]?
     public var isBase64Encoded: Bool?
 
     public init(
-        statusCode: HTTPResponseStatus,
+        statusCode: HTTPResponse.Status,
         headers: HTTPHeaders? = nil,
         body: String? = nil,
         cookies: [String]? = nil,
@@ -101,12 +106,3 @@ public struct FunctionURLResponse: Codable {
         self.isBase64Encoded = isBase64Encoded
     }
 }
-
-#if swift(>=5.6)
-extension FunctionURLRequest: Sendable {}
-extension FunctionURLRequest.Context: Sendable {}
-extension FunctionURLRequest.Context.Authorizer: Sendable {}
-extension FunctionURLRequest.Context.Authorizer.IAMAuthorizer: Sendable {}
-extension FunctionURLRequest.Context.HTTP: Sendable {}
-extension FunctionURLResponse: Sendable {}
-#endif
