@@ -12,12 +12,16 @@
 //
 //===----------------------------------------------------------------------===//
 
-import struct Foundation.Date
+#if canImport(FoundationEssentials)
+import FoundationEssentials
+#else
+import Foundation
+#endif
 
 // https://docs.aws.amazon.com/lambda/latest/dg/with-s3.html
 
-public struct S3Event: Decodable {
-    public struct Record: Decodable {
+public struct S3Event: Decodable, Sendable {
+    public struct Record: Decodable, Sendable {
         public let eventVersion: String
         public let eventSource: String
         public let awsRegion: AWSRegion
@@ -37,15 +41,15 @@ public struct S3Event: Decodable {
         case records = "Records"
     }
 
-    public struct RequestParameters: Codable, Equatable {
+    public struct RequestParameters: Codable, Equatable, Sendable {
         public let sourceIPAddress: String
     }
 
-    public struct UserIdentity: Codable, Equatable {
+    public struct UserIdentity: Codable, Equatable, Sendable {
         public let principalId: String
     }
 
-    public struct Entity: Codable {
+    public struct Entity: Codable, Sendable {
         public let configurationId: String
         public let schemaVersion: String
         public let bucket: Bucket
@@ -59,13 +63,13 @@ public struct S3Event: Decodable {
         }
     }
 
-    public struct Bucket: Codable {
+    public struct Bucket: Codable, Sendable {
         public let name: String
         public let ownerIdentity: UserIdentity
         public let arn: String
     }
 
-    public struct Object: Codable {
+    public struct Object: Codable, Sendable {
         public let key: String
         /// The object's size in bytes.
         ///
@@ -73,17 +77,7 @@ public struct S3Event: Decodable {
         public let size: UInt64?
         public let urlDecodedKey: String?
         public let versionId: String?
-        public let eTag: String
+        public let eTag: String?
         public let sequencer: String
     }
 }
-
-#if swift(>=5.6)
-extension S3Event: Sendable {}
-extension S3Event.Bucket: Sendable {}
-extension S3Event.Entity: Sendable {}
-extension S3Event.Object: Sendable {}
-extension S3Event.Record: Sendable {}
-extension S3Event.RequestParameters: Sendable {}
-extension S3Event.UserIdentity: Sendable {}
-#endif
