@@ -36,11 +36,14 @@ extension Base64 {
     }
 
     @inlinable
-    static func decode<Buffer: Collection>(encoded: Buffer, options: DecodingOptions = [])
-        throws -> [UInt8] where Buffer.Element == UInt8 {
-        let alphabet = options.contains(.base64UrlAlphabet)
-            ? Base64.decodeBase64Url
-            : Base64.decodeBase64
+    static func decode<Buffer: Collection>(
+        encoded: Buffer,
+        options: DecodingOptions = []
+    ) throws -> [UInt8] where Buffer.Element == UInt8 {
+        let alphabet =
+            options.contains(.base64UrlAlphabet)
+            ? Base64.urlAlphabet
+            : Base64.defaultAlphabet
 
         // In Base64 4 encoded bytes, become 3 decoded bytes. We pad to the
         // nearest multiple of three.
@@ -58,7 +61,7 @@ extension Base64 {
         outputBytes.reserveCapacity(outputLength)
 
         // fast loop. we don't expect any padding in here.
-        for _ in 0 ..< fullQualified {
+        for _ in 0..<fullQualified {
             let firstValue: UInt8 = try iterator.nextValue(alphabet: alphabet)
             let secondValue: UInt8 = try iterator.nextValue(alphabet: alphabet)
             let thirdValue: UInt8 = try iterator.nextValue(alphabet: alphabet)
@@ -109,65 +112,65 @@ extension Base64 {
     // MARK: Internal
 
     @usableFromInline
-    static let decodeBase64: [UInt8] = [
+    static let defaultAlphabet: [UInt8] = [
         //     0    1    2    3    4    5    6    7    8    9
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, //  0
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, //  1
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, //  2
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, //  3
-        255, 255, 255, 62, 255, 255, 255, 63, 52, 53, //  4
-        54, 55, 56, 57, 58, 59, 60, 61, 255, 255, //  5
-        255, 254, 255, 255, 255, 0, 1, 2, 3, 4, //  6
-        5, 6, 7, 8, 9, 10, 11, 12, 13, 14, //  7
-        15, 16, 17, 18, 19, 20, 21, 22, 23, 24, //  8
-        25, 255, 255, 255, 255, 255, 255, 26, 27, 28, //  9
-        29, 30, 31, 32, 33, 34, 35, 36, 37, 38, // 10
-        39, 40, 41, 42, 43, 44, 45, 46, 47, 48, // 11
-        49, 50, 51, 255, 255, 255, 255, 255, 255, 255, // 12
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 13
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 14
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 15
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 16
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 17
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 18
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 19
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 20
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 21
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 22
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 23
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 24
-        255, 255, 255, 255, 255, // 25
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  //  0
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  //  1
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  //  2
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  //  3
+        255, 255, 255, 62, 255, 255, 255, 63, 52, 53,  //  4
+        54, 55, 56, 57, 58, 59, 60, 61, 255, 255,  //  5
+        255, 254, 255, 255, 255, 0, 1, 2, 3, 4,  //  6
+        5, 6, 7, 8, 9, 10, 11, 12, 13, 14,  //  7
+        15, 16, 17, 18, 19, 20, 21, 22, 23, 24,  //  8
+        25, 255, 255, 255, 255, 255, 255, 26, 27, 28,  //  9
+        29, 30, 31, 32, 33, 34, 35, 36, 37, 38,  // 10
+        39, 40, 41, 42, 43, 44, 45, 46, 47, 48,  // 11
+        49, 50, 51, 255, 255, 255, 255, 255, 255, 255,  // 12
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 13
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 14
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 15
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 16
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 17
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 18
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 19
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 20
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 21
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 22
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 23
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 24
+        255, 255, 255, 255, 255,  // 25
     ]
 
     @usableFromInline
-    static let decodeBase64Url: [UInt8] = [
+    static let urlAlphabet: [UInt8] = [
         //     0    1    2    3    4    5    6    7    8    9
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, //  0
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, //  1
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, //  2
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, //  3
-        255, 255, 255, 255, 255, 62, 255, 255, 52, 53, //  4
-        54, 55, 56, 57, 58, 59, 60, 61, 255, 255, //  5
-        255, 254, 255, 255, 255, 0, 1, 2, 3, 4, //  6
-        5, 6, 7, 8, 9, 10, 11, 12, 13, 14, //  7
-        15, 16, 17, 18, 19, 20, 21, 22, 23, 24, //  8
-        25, 255, 255, 255, 255, 63, 255, 26, 27, 28, //  9
-        29, 30, 31, 32, 33, 34, 35, 36, 37, 38, // 10
-        39, 40, 41, 42, 43, 44, 45, 46, 47, 48, // 11
-        49, 50, 51, 255, 255, 255, 255, 255, 255, 255, // 12
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 13
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 14
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 15
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 16
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 17
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 18
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 19
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 20
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 21
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 22
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 23
-        255, 255, 255, 255, 255, 255, 255, 255, 255, 255, // 24
-        255, 255, 255, 255, 255, // 25
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  //  0
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  //  1
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  //  2
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  //  3
+        255, 255, 255, 255, 255, 62, 255, 255, 52, 53,  //  4
+        54, 55, 56, 57, 58, 59, 60, 61, 255, 255,  //  5
+        255, 254, 255, 255, 255, 0, 1, 2, 3, 4,  //  6
+        5, 6, 7, 8, 9, 10, 11, 12, 13, 14,  //  7
+        15, 16, 17, 18, 19, 20, 21, 22, 23, 24,  //  8
+        25, 255, 255, 255, 255, 63, 255, 26, 27, 28,  //  9
+        29, 30, 31, 32, 33, 34, 35, 36, 37, 38,  // 10
+        39, 40, 41, 42, 43, 44, 45, 46, 47, 48,  // 11
+        49, 50, 51, 255, 255, 255, 255, 255, 255, 255,  // 12
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 13
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 14
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 15
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 16
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 17
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 18
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 19
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 20
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 21
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 22
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 23
+        255, 255, 255, 255, 255, 255, 255, 255, 255, 255,  // 24
+        255, 255, 255, 255, 255,  // 25
     ]
 
     @usableFromInline
