@@ -12,91 +12,92 @@
 //
 //===----------------------------------------------------------------------===//
 
-@testable import AWSLambdaEvents
 import XCTest
+
+@testable import AWSLambdaEvents
 
 class S3Tests: XCTestCase {
     static let eventBodyObjectCreated = """
-    {
-      "Records": [
         {
-          "eventVersion":"2.1",
-          "eventSource":"aws:s3",
-          "awsRegion":"eu-central-1",
-          "eventTime":"2020-01-13T09:25:40.621Z",
-          "eventName":"ObjectCreated:Put",
-          "userIdentity":{
-            "principalId":"AWS:AAAAAAAJ2MQ4YFQZ7AULJ"
-          },
-          "requestParameters":{
-            "sourceIPAddress":"123.123.123.123"
-          },
-          "responseElements":{
-            "x-amz-request-id":"01AFA1430E18C358",
-            "x-amz-id-2":"JsbNw6sHGFwgzguQjbYcew//bfAeZITyTYLfjuu1U4QYqCq5CPlSyYLtvWQS+gw0RxcroItGwm8="
-          },
-          "s3":{
-            "s3SchemaVersion":"1.0",
-            "configurationId":"98b55bc4-3c0c-4007-b727-c6b77a259dde",
-            "bucket":{
-              "name":"eventsources",
-              "ownerIdentity":{
-                "principalId":"AAAAAAAAAAAAAA"
+          "Records": [
+            {
+              "eventVersion":"2.1",
+              "eventSource":"aws:s3",
+              "awsRegion":"eu-central-1",
+              "eventTime":"2020-01-13T09:25:40.621Z",
+              "eventName":"ObjectCreated:Put",
+              "userIdentity":{
+                "principalId":"AWS:AAAAAAAJ2MQ4YFQZ7AULJ"
               },
-              "arn":"arn:aws:s3:::eventsources"
-            },
-            "object":{
-              "key":"Hi.md",
-              "size":2880,
-              "eTag":"91a7f2c3ae81bcc6afef83979b463f0e",
-              "sequencer":"005E1C37948E783A6E"
+              "requestParameters":{
+                "sourceIPAddress":"123.123.123.123"
+              },
+              "responseElements":{
+                "x-amz-request-id":"01AFA1430E18C358",
+                "x-amz-id-2":"JsbNw6sHGFwgzguQjbYcew//bfAeZITyTYLfjuu1U4QYqCq5CPlSyYLtvWQS+gw0RxcroItGwm8="
+              },
+              "s3":{
+                "s3SchemaVersion":"1.0",
+                "configurationId":"98b55bc4-3c0c-4007-b727-c6b77a259dde",
+                "bucket":{
+                  "name":"eventsources",
+                  "ownerIdentity":{
+                    "principalId":"AAAAAAAAAAAAAA"
+                  },
+                  "arn":"arn:aws:s3:::eventsources"
+                },
+                "object":{
+                  "key":"Hi.md",
+                  "size":2880,
+                  "eTag":"91a7f2c3ae81bcc6afef83979b463f0e",
+                  "sequencer":"005E1C37948E783A6E"
+                }
+              }
             }
-          }
+          ]
         }
-      ]
-    }
-    """
+        """
 
     // A S3 ObjectRemoved:* event does not contain the object size
     static let eventBodyObjectRemoved = """
-    {
-      "Records": [
         {
-          "eventVersion":"2.1",
-          "eventSource":"aws:s3",
-          "awsRegion":"eu-central-1",
-          "eventTime":"2020-01-13T09:25:40.621Z",
-          "eventName":"ObjectRemoved:DeleteMarkerCreated",
-          "userIdentity":{
-            "principalId":"AWS:AAAAAAAJ2MQ4YFQZ7AULJ"
-          },
-          "requestParameters":{
-            "sourceIPAddress":"123.123.123.123"
-          },
-          "responseElements":{
-            "x-amz-request-id":"01AFA1430E18C358",
-            "x-amz-id-2":"JsbNw6sHGFwgzguQjbYcew//bfAeZITyTYLfjuu1U4QYqCq5CPlSyYLtvWQS+gw0RxcroItGwm8="
-          },
-          "s3":{
-            "s3SchemaVersion":"1.0",
-            "configurationId":"98b55bc4-3c0c-4007-b727-c6b77a259dde",
-            "bucket":{
-              "name":"eventsources",
-              "ownerIdentity":{
-                "principalId":"AAAAAAAAAAAAAA"
+          "Records": [
+            {
+              "eventVersion":"2.1",
+              "eventSource":"aws:s3",
+              "awsRegion":"eu-central-1",
+              "eventTime":"2020-01-13T09:25:40.621Z",
+              "eventName":"ObjectRemoved:DeleteMarkerCreated",
+              "userIdentity":{
+                "principalId":"AWS:AAAAAAAJ2MQ4YFQZ7AULJ"
               },
-              "arn":"arn:aws:s3:::eventsources"
-            },
-            "object":{
-              "key":"Hi.md",
-              "eTag":"91a7f2c3ae81bcc6afef83979b463f0e",
-              "sequencer":"005E1C37948E783A6E"
+              "requestParameters":{
+                "sourceIPAddress":"123.123.123.123"
+              },
+              "responseElements":{
+                "x-amz-request-id":"01AFA1430E18C358",
+                "x-amz-id-2":"JsbNw6sHGFwgzguQjbYcew//bfAeZITyTYLfjuu1U4QYqCq5CPlSyYLtvWQS+gw0RxcroItGwm8="
+              },
+              "s3":{
+                "s3SchemaVersion":"1.0",
+                "configurationId":"98b55bc4-3c0c-4007-b727-c6b77a259dde",
+                "bucket":{
+                  "name":"eventsources",
+                  "ownerIdentity":{
+                    "principalId":"AAAAAAAAAAAAAA"
+                  },
+                  "arn":"arn:aws:s3:::eventsources"
+                },
+                "object":{
+                  "key":"Hi.md",
+                  "eTag":"91a7f2c3ae81bcc6afef83979b463f0e",
+                  "sequencer":"005E1C37948E783A6E"
+                }
+              }
             }
-          }
+          ]
         }
-      ]
-    }
-    """
+        """
 
     func testSimpleEventFromJSON() {
         let data = S3Tests.eventBodyObjectCreated.data(using: .utf8)!
