@@ -12,8 +12,9 @@
 //
 //===----------------------------------------------------------------------===//
 
-@testable import AWSLambdaEvents
 import XCTest
+
+@testable import AWSLambdaEvents
 
 class Base64Tests: XCTestCase {
     // MARK: - Decoding -
@@ -32,9 +33,10 @@ class Base64Tests: XCTestCase {
     }
 
     func testBase64DecodingAllTheBytesSequentially() {
-        let base64 = "AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8gISIjJCUmJygpKissLS4vMDEyMzQ1Njc4OTo7PD0+P0BBQkNERUZHSElKS0xNTk9QUVJTVFVWV1hZWltcXV5fYGFiY2RlZmdoaWprbG1ub3BxcnN0dXZ3eHl6e3x9fn+AgYKDhIWGh4iJiouMjY6PkJGSk5SVlpeYmZqbnJ2en6ChoqOkpaanqKmqq6ytrq+wsbKztLW2t7i5uru8vb6/wMHCw8TFxsfIycrLzM3Oz9DR0tPU1dbX2Nna29zd3t/g4eLj5OXm5+jp6uvs7e7v8PHy8/T19vf4+fr7/P3+/w=="
+        let base64 =
+            "AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8gISIjJCUmJygpKissLS4vMDEyMzQ1Njc4OTo7PD0+P0BBQkNERUZHSElKS0xNTk9QUVJTVFVWV1hZWltcXV5fYGFiY2RlZmdoaWprbG1ub3BxcnN0dXZ3eHl6e3x9fn+AgYKDhIWGh4iJiouMjY6PkJGSk5SVlpeYmZqbnJ2en6ChoqOkpaanqKmqq6ytrq+wsbKztLW2t7i5uru8vb6/wMHCw8TFxsfIycrLzM3Oz9DR0tPU1dbX2Nna29zd3t/g4eLj5OXm5+jp6uvs7e7v8PHy8/T19vf4+fr7/P3+/w=="
 
-        let expected = Array(UInt8(0) ... UInt8(255))
+        let expected = Array(UInt8(0)...UInt8(255))
         var decoded: [UInt8]?
         XCTAssertNoThrow(decoded = try base64.base64decoded())
 
@@ -42,9 +44,10 @@ class Base64Tests: XCTestCase {
     }
 
     func testBase64UrlDecodingAllTheBytesSequentially() {
-        let base64 = "AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8gISIjJCUmJygpKissLS4vMDEyMzQ1Njc4OTo7PD0-P0BBQkNERUZHSElKS0xNTk9QUVJTVFVWV1hZWltcXV5fYGFiY2RlZmdoaWprbG1ub3BxcnN0dXZ3eHl6e3x9fn-AgYKDhIWGh4iJiouMjY6PkJGSk5SVlpeYmZqbnJ2en6ChoqOkpaanqKmqq6ytrq-wsbKztLW2t7i5uru8vb6_wMHCw8TFxsfIycrLzM3Oz9DR0tPU1dbX2Nna29zd3t_g4eLj5OXm5-jp6uvs7e7v8PHy8_T19vf4-fr7_P3-_w=="
+        let base64 =
+            "AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8gISIjJCUmJygpKissLS4vMDEyMzQ1Njc4OTo7PD0-P0BBQkNERUZHSElKS0xNTk9QUVJTVFVWV1hZWltcXV5fYGFiY2RlZmdoaWprbG1ub3BxcnN0dXZ3eHl6e3x9fn-AgYKDhIWGh4iJiouMjY6PkJGSk5SVlpeYmZqbnJ2en6ChoqOkpaanqKmqq6ytrq-wsbKztLW2t7i5uru8vb6_wMHCw8TFxsfIycrLzM3Oz9DR0tPU1dbX2Nna29zd3t_g4eLj5OXm5-jp6uvs7e7v8PHy8_T19vf4-fr7_P3-_w=="
 
-        let expected = Array(UInt8(0) ... UInt8(255))
+        let expected = Array(UInt8(0)...UInt8(255))
         var decoded: [UInt8]?
         XCTAssertNoThrow(decoded = try base64.base64decoded(options: .base64UrlAlphabet))
 
@@ -52,13 +55,13 @@ class Base64Tests: XCTestCase {
     }
 
     func testBase64DecodingWithPoop() {
-        XCTAssertThrowsError(_ = try "💩".base64decoded()) { error in
+        XCTAssertThrowsError(try "💩".base64decoded()) { error in
             XCTAssertEqual(error as? Base64.DecodingError, .invalidCharacter(240))
         }
     }
 
     func testBase64DecodingWithInvalidLength() {
-        XCTAssertThrowsError(_ = try "AAAAA".base64decoded()) { error in
+        XCTAssertThrowsError(try "AAAAA".base64decoded()) { error in
             XCTAssertEqual(error as? Base64.DecodingError, .invalidLength)
         }
     }
